+++ apps/wallets/smart-wallet/next/src/app/page.tsx
@@ -1,26 +1,24 @@
 "use client";
 
-import { GoogleSignInButton } from "@/components/google-signin-button";
+import { Fireworks } from "@/components/fireworks";
 import { MintNFTButton } from "@/components/mint-nft-button";
 import { PoweredByCrossmint } from "@/components/powered-by-crossmint";
+import { SignInAuthButton } from "@/components/signin-auth-button";
 import { Typography } from "@/components/typography";
-import { useToast } from "@/components/use-toast";
 import Link from "next/link";
+import { useState } from "react";
 
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "@crossmint/client-sdk-react-ui";
 
 export default function Home() {
-    const { isLoading, authedUser } = useAuth();
-    const { toasts } = useToast();
+    const { jwt, wallet } = useAuth();
+    const [nftSuccessfullyMinted, setNftSuccessfullyMinted] = useState(false);
 
-    const showMintButton = !isLoading && authedUser;
-
-    const nftSuccessfullyMinted = !!toasts.find((toast) => toast.title?.includes("NFT Minted"));
+    const showMintButton = jwt != null && wallet != null;
 
     return (
-        <div className="flex h-full w-full items-center pt-6 justify-center">
-            <div className="flex flex-col pb-12 items-center max-w-[505px] p-4">
-                {/* todo: Only show the demo verbiage if nft hasn't been minted. */}
+        <div className="flex h-full w-full items-center md:p-4 justify-center">
+            <div className="flex flex-col pb-12 items-center max-w-[538px] p-4">
                 <div className="flex flex-col gap-2 text-center pb-8">
                     <Typography
                         style={{
@@ -39,7 +37,7 @@ export default function Home() {
                 </div>
 
                 <div className="flex flex-col w-full md:max-w-[340px] gap-10">
-                    <div className="bg-card flex flex-col  p-5 rounded-3xl shadow-dropdown">
+                    <div className="bg-card flex flex-col p-5 rounded-3xl shadow-dropdown">
                         {/* eslint-disable-next-line @next/next/no-img-element */}
                         <img className="rounded-xl rounded-bl-none rounded-br-none" src={"/emoji-nft.png"} alt="nft" />
                         <div className="py-4">
@@ -51,9 +49,10 @@ export default function Home() {
                             </Typography>
                         </div>
                     </div>
+                    <Fireworks play={nftSuccessfullyMinted} />
 
                     {showMintButton && nftSuccessfullyMinted ? (
-                        <div className="flex gap-2 items-center self-center min-h-12">
+                        <div className="flex gap-2 items-center self-center min-h-[52px]">
                             <Link
                                 href="/wallet"
                                 className="underline text-secondary-foreground text-lg font-semibold underline-offset-4"
@@ -62,8 +61,10 @@ export default function Home() {
                             </Link>
                         </div>
                     ) : null}
-                    {showMintButton && !nftSuccessfullyMinted ? <MintNFTButton /> : null}
-                    {!showMintButton && !nftSuccessfullyMinted ? <GoogleSignInButton /> : null}
+                    {showMintButton && !nftSuccessfullyMinted ? (
+                        <MintNFTButton setNftSuccessfullyMinted={setNftSuccessfullyMinted} />
+                    ) : null}
+                    {!showMintButton && !nftSuccessfullyMinted ? <SignInAuthButton /> : null}
 
                     <PoweredByCrossmint />
                 </div>