+++ packages/client/auth-core/src/contexts/authlogin.tsx
@@ -17,13 +17,21 @@ const AuthContext = createContext<AuthContextType>({
     jwt: null,
 });
 
-type AuthProviderParams = {
+export type AuthProviderParams = {
     apiKey: string;
     children: ReactNode;
 };
 
+const getJwtFromCookie = (): string | null => {
+    if (typeof document === "undefined") {
+        return null; // Check if we're on the client-side
+    }
+    const crossmintSession = document.cookie.split("; ").find((row) => row.startsWith("crossmint-session"));
+    return crossmintSession ? crossmintSession.split("=")[1] : null;
+};
+
 export function AuthProvider({ children, apiKey }: AuthProviderParams) {
-    const [jwtToken, setJwtToken] = useState<string | null>(null);
+    const [jwtToken, setJwtToken] = useState<string | null>(() => getJwtFromCookie());
     const [modalOpen, setModalOpen] = useState(false);
     const crossmintService = useMemo(() => new CrossmintService(apiKey, jwtToken), [apiKey, jwtToken]);