+++ packages/client/wallets/smart-wallet/src/api/CrossmintWalletService.ts
@@ -1,19 +1,20 @@
-import { SmartWalletChain } from "@/blockchain/chains";
-import { SignerData, StoreSmartWalletParams } from "@/types/API";
-import type { UserParams } from "@/types/Config";
-import { API_VERSION } from "@/utils/constants";
-import { UserOperation } from "permissionless";
-import { GetEntryPointVersion } from "permissionless/_types/types";
-import { EntryPoint } from "permissionless/types/entrypoint";
+import type { UserOperation } from "permissionless";
+import type { EntryPoint, GetEntryPointVersion } from "permissionless/types/entrypoint";
 
+import { CrossmintServiceError } from "@crossmint/client-sdk-base";
 import { blockchainToChainId } from "@crossmint/common-sdk-base";
 
+import { BaseCrossmintService } from "../api/BaseCrossmintService";
+import type { SmartWalletChain } from "../blockchain/chains";
+import type { UserParams } from "../types/params";
+import { SmartWalletConfigSchema } from "../types/schema";
+import type { SmartWalletConfig, StoreSmartWalletParams } from "../types/service";
 import { bigintsToHex, parseBigintAPIResponse } from "../utils/api";
-import { BaseCrossmintService } from "./BaseCrossmintService";
+import { API_VERSION } from "../utils/constants";
 
 export class CrossmintWalletService extends BaseCrossmintService {
-    async idempotentCreateSmartWallet(user: UserParams, input: StoreSmartWalletParams) {
-        return this.fetchCrossmintAPI(
+    async idempotentCreateSmartWallet(user: UserParams, input: StoreSmartWalletParams): Promise<void> {
+        await this.fetchCrossmintAPI(
             `${API_VERSION}/sdk/smart-wallet`,
             { method: "PUT", body: JSON.stringify(input) },
             "Error creating abstract wallet. Please contact support",
@@ -37,22 +38,22 @@ export class CrossmintWalletService extends BaseCrossmintService {
         return parseBigintAPIResponse(result);
     }
 
-    async getSmartWalletConfig(
-        user: UserParams,
-        chain: SmartWalletChain
-    ): Promise<{
-        kernelVersion: string;
-        entryPointVersion: string;
-        userId: string;
-        signers: { signerData: SignerData }[];
-        smartContractWalletAddress?: string;
-    }> {
-        return this.fetchCrossmintAPI(
+    async getSmartWalletConfig(user: UserParams, chain: SmartWalletChain): Promise<SmartWalletConfig> {
+        const data: unknown = await this.fetchCrossmintAPI(
             `${API_VERSION}/sdk/smart-wallet/config?chain=${chain}`,
             { method: "GET" },
             "Error getting smart wallet version configuration. Please contact support",
             user.jwt
         );
+
+        const result = SmartWalletConfigSchema.safeParse(data);
+        if (!result.success) {
+            throw new CrossmintServiceError(
+                `Invalid smart wallet config, please contact support. Details below:\n${result.error.toString()}`
+            );
+        }
+
+        return result.data;
     }
 
     async fetchNFTs(address: string, chain: SmartWalletChain) {