+++ packages/client/wallets/smart-wallet/src/blockchain/wallets/EVMSmartWallet.ts
@@ -1,10 +1,11 @@
-import { logError, logInfo } from "@/services/logging";
 import { type HttpTransport, type PublicClient, isAddress, publicActions } from "viem";
 
+import { TransferError } from "@crossmint/client-sdk-base";
+
 import type { CrossmintWalletService } from "../../api/CrossmintWalletService";
-import { TransferError } from "../../error";
-import type { TransferType } from "../../types/Tokens";
+import { logError, logInfo } from "../../services/logging";
 import { SmartWalletClient } from "../../types/internal";
+import type { TransferType } from "../../types/token";
 import { SCW_SERVICE } from "../../utils/constants";
 import { errorToJSON, logPerformance } from "../../utils/log";
 import { SmartWalletChain } from "../chains";