+++ packages/client/wallets/smart-wallet/src/blockchain/wallets/account/strategy.ts
@@ -1,0 +1,5 @@
+import { AccountAndSigner, WalletCreationContext } from "../../../types/internal";
+
+export interface AccountCreationStrategy {
+    create(params: WalletCreationContext): Promise<AccountAndSigner>;
+}