+++ apps/payments/nextjs/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @crossmint/client-sdk-nextjs-starter
 
+## 1.2.6
+
+### Patch Changes
+
+-   @crossmint/client-sdk-react-ui@1.3.3
+
+## 1.2.5
+
+### Patch Changes
+
+-   Updated dependencies [5c60365]
+    -   @crossmint/client-sdk-react-ui@1.3.2
+
+## 1.2.4
+
+### Patch Changes
+
+-   Updated dependencies [12b5eac]
+    -   @crossmint/client-sdk-base@1.2.2
+    -   @crossmint/client-sdk-react-ui@1.3.1
+
+## 1.2.3
+
+### Patch Changes
+
+-   Updated dependencies [71b3778]
+    -   @crossmint/client-sdk-react-ui@1.3.0
+
 ## 1.2.2
 
 ### Patch Changes