+++ packages/client/auth-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @crossmint/client-sdk-auth-core
 
+## 1.1.1
+
+### Patch Changes
+
+-   e9060b6: Consolidated exports and improved initial jwt state
+
+## 1.1.0
+
+### Minor Changes
+
+-   dc97a9a: Change query parameter
+
 ## 1.0.2
 
 ### Patch Changes